requirements.txt
fastapi
uvicorn
python-dotenv
serpapi
python-multipart